requirements.txt
ollama
SpeechRecognition
sounddevice
pyaudio 
numpy
soundfile
torch
openai-whisper
python-dotenv
langchain_community
playwright
beautifulsoup4

langchain 
langchain-community 
langchainhub 
langchain-openai 
chromadb 
bs4
faiss-cpu